requirements.txt
flask==3.0.2
flask-socketio==5.3.6
pillow==10.2.0
numpy==1.26.4
mss==9.0.1
ollama==0.1.14
requests==2.31.0
twilio==8.12.0
python-dotenv==1.0.1
pynput==1.7.6
customtkinter==5.2.2